New clippy warnings.
Don't want CI to fail on stricter professional setups, so commit these
changes to use unwrap_or_default.

To downstream, don't worry, it is not a security release.

Incentives are funny things, aren't they?  In the push to safe
languages, a nice flywheel is set in motion to build a future ontology
and corresponding custodian roles for security release tags and
release subscription models.

Golden Rule: hopefully it is a fun set of projects.