Start seriously implementing Commodore D64 disks.
This format is behind a feature flag named "commodore" for now.
So it won't pull in the forbidden-bands dependency unless you need
D64 support.

Refactored configuration throughout the library, bumped to version
0.8.0 because it is incompatible.

Added a "catalog" trait to list the contents of a disk.  I still need
to figure out a good API for this crate, but it should be helpful to
others who just want to engage in some fun software archaeology.