Start working on the API refactor for DiskImage traits.
At lot of this work will be undone in the next couple commits, but it
gets some of the basic patterns out there.

This adds a DiskImageGuesser trait, which lets you use heuristics to
identify the probable type of a disk image from a filename and byte
slice.

It also provides a parse function that lets you parse a
DiskImageGuess.  This trait will probably be split up.